Fix circular import of PipelineRunConfiguration
