+++ src/zenml/config/pipeline_configurations.py
@@ -12,18 +12,13 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 """Pipeline configuration classes."""
-from typing import TYPE_CHECKING, Any, Dict, Optional, Union
-from uuid import UUID
+from typing import TYPE_CHECKING, Any, Dict, Optional
 
 from pydantic import validator
 
 from zenml.config.constants import DOCKER_SETTINGS_KEY
-from zenml.config.schedule import Schedule
 from zenml.config.source import Source, convert_source_validator
-from zenml.config.step_configurations import StepConfigurationUpdate
 from zenml.config.strict_base_model import StrictBaseModel
-from zenml.models.pipeline_build_models import PipelineBuildBaseModel
-from zenml.utils import pydantic_utils
 
 if TYPE_CHECKING:
     from zenml.config import DockerSettings
@@ -86,18 +81,3 @@ class PipelineConfiguration(PipelineConfigurationUpdate):
             DOCKER_SETTINGS_KEY, {}
         )
         return DockerSettings.parse_obj(model_or_dict)
-
-
-class PipelineRunConfiguration(
-    StrictBaseModel, pydantic_utils.YAMLSerializationMixin
-):
-    """Class for pipeline run configurations."""
-
-    run_name: Optional[str] = None
-    enable_cache: Optional[bool] = None
-    enable_artifact_metadata: Optional[bool] = None
-    schedule: Optional[Schedule] = None
-    build: Union[PipelineBuildBaseModel, UUID, None] = None
-    steps: Dict[str, StepConfigurationUpdate] = {}
-    settings: Dict[str, BaseSettings] = {}
-    extra: Dict[str, Any] = {}
